+++ test/unit-test/test_cli_arg.py
@@ -605,10 +605,14 @@ def test_list_deployments_extra_args():
 @pytest.mark.list_deployments
 def test_list_deployments(session_manager, mocker):
     mocker.patch("seedfarmer.cli_groups._list_group.mi.get_all_deployments", return_value=None)
+    mocker.patch(
+        "seedfarmer.cli_groups._list_group.get_sts_identity_info",
+        return_value=("1234566789012", "arn:aws", "aws"),
+    )
     _test_command(
         sub_command=list,
         options=["deployments", "-p", "myapp", "--debug"],
-        exit_code=1,
+        exit_code=0,
     )